Rename mod and update descriptions internally
