+++ experiments/minecraft/hifimc/src/main/java/com/highfidelity/hifimc/StaticHttpHandler.java
@@ -1,4 +1,4 @@
-package com.example.examplemod;
+package com.highfidelity.hifimc;
 
 import java.io.IOException;
 import java.io.PrintWriter;
